wip ifc subprocess rework
